Untested ExtrapolationConstrainedOptimizer
