+++ cooper/optim/constrained_optimizers/constrained_optimizer.py
@@ -3,8 +3,7 @@
 Implementation of the :py:class:`ConstrainedOptimizer` class.
 """
 
-from collections.abc import Sequence
-from typing import Optional, Union
+from typing import Optional
 
 import torch
 
@@ -98,7 +97,7 @@ class ConstrainedOptimizer:
         for dual_optimizer in self.dual_optimizers:
             dual_optimizer.zero_grad()
 
-    def dual_step(self, call_extrapolation: bool = False):
+    def dual_step(self):
         """
         Perform a gradient step on the parameters associated with the dual variables.
         Since the dual problem involves *maximizing* over the dual variables, we flip
@@ -107,12 +106,6 @@ class ConstrainedOptimizer:
         After being updated by the dual optimizer steps, the multipliers are
         post-processed (e.g. to ensure feasibility for equality constraints, or to
         apply dual restarts).
-
-        Args:
-            call_extrapolation: Whether to call ``dual_optimizer.extrapolation()`` as
-                opposed to ``dual_optimizer.step()``. This is only relevant for
-                :py:class:`~cooper.optim.constrained_optimizers.ExtrapolationConstrainedOptimizer`
-                and should be left to ``False`` for other ``ConstrainedOptimizer``\\s.
         """
         for multiplier in self.multipliers:
             for param in multiplier.parameters():
@@ -125,11 +118,7 @@ class ConstrainedOptimizer:
         for dual_optimizer in self.dual_optimizers:
             # Update multipliers based on current constraint violations (gradients)
             # For unobserved constraints the gradient is None, so this is a no-op.
-            if call_extrapolation:
-                assert self.extrapolation
-                dual_optimizer.extrapolation()
-            else:
-                dual_optimizer.step()
+            dual_optimizer.step()
 
         for multiplier in self.multipliers:
             if isinstance(multiplier, ExplicitMultiplier):
